Create model for representing a tree-scheme.
Code is slighly verbose but used the same pattern i've also used for the tree.

The scheme itself is immutable and you create a new scheme through a
builder object. Additional advantage of this is that we allready can add
many constraints to the building of a scheme so we are guaranteed when
we have a scheme object that it is well formed.