Update array order buttons to cycle
If pressing 'up' on the first element it used to be swapped with the
the last element, this is however very confusing behaviour. With this
fix the element will be 'moved' to the end preserving the order of the
rest of the array. Same story for the 'down' button on the last element.